refactor(pickers): move projected toggle handling to base directive
